+++ main_project/component/component.hpp
@@ -20,20 +20,20 @@ public:
     virtual ~Component() = 0; //destructor of base class should be virtual (purely virtual as component should be abstract)
 
     virtual double getConductance() const;
-    virtual float getVoltage() const;
+    virtual double getVoltage() const;
     virtual double getCurrent() const; //For complex components, this returns the current through the companion model's current source rather than through the whole component
-    virtual double getTotalCurrent(float voltage, int order = 1); //For complex components, this return the current through the whole component
+    virtual double getTotalCurrent(double voltage, int order = 1); //For complex components, this return the current through the whole component
 
     string getName() const;
     virtual vector<int> getNodes() const = 0;
 
 
     // this should be used to update the value of the voltage and current accross a component after an iteration
-    virtual void updateVals(float newVoltage, float newCurrent, int order);
+    virtual void updateVals(double newVoltage, double newCurrent, int order);
     // this should be used to update the time of the time dependant components or update nonlinear components (passing voltage)
-    virtual void updateVals(float time_or_voltage);
+    virtual void updateVals(double time_or_voltage);
 
-    float getValue(string val) const;
+    double getValue(string val) const;
     vector<componentType> getTypes() const;
 };
