+++ main_project/component/currentSource.cpp
@@ -3,7 +3,7 @@
 
 #include "currentSource.hpp"
 
-CurrentSource::CurrentSource(string name, vector<string> args, vector<float> extraInfo)
+CurrentSource::CurrentSource(string name, vector<string> args, vector<double> extraInfo)
     :Component{name}
 {
     int n1 = stoi(args[0]);
@@ -12,7 +12,7 @@ CurrentSource::CurrentSource(string name, vector<string> args, vector<float> ext
     setupBasic(n1, n2);
 
     if(args.size() == 3){
-        float val = getValue(args[2]);
+        double val = getValue(args[2]);
         setupDC(val);
     }else{
         string flow = args[2];
@@ -27,7 +27,7 @@ CurrentSource::CurrentSource(string name, vector<string> args, vector<float> ext
     }
 }
 
-CurrentSource::CurrentSource(string _name, float _current, int n1, int n2)
+CurrentSource::CurrentSource(string _name, double _current, int n1, int n2)
     :Component{_name}{
     setupBasic(n1, n2);
     setupDC(_current);
@@ -39,7 +39,7 @@ void CurrentSource::setupBasic(int n1, int n2){
     types.push_back(componentType::currentSource);
 }
 
-void CurrentSource::setupDC(float _current){
+void CurrentSource::setupDC(double _current){
     current = _current;
 }
 
@@ -47,7 +47,7 @@ double CurrentSource::getCurrent() const{
     return current;
 }
 
-void CurrentSource::updateVals(float time){
+void CurrentSource::updateVals(double time){
     current = currentWaveform.updateVals(time);
 }
