+++ main_project/component/resistor.cpp
@@ -3,12 +3,12 @@
 
 #include "resistor.hpp"
 
-Resistor::Resistor(string name, vector<string> args, vector<float> extraInfo)
+Resistor::Resistor(string name, vector<string> args, vector<double> extraInfo)
     :Component{name}
 {
     int n1 = stoi(args[0]);
     int n2 = stoi(args[1]);
-    float val = getValue(args[2]);
+    double val = getValue(args[2]);
 
     nodes.push_back(n1);
     nodes.push_back(n2);
@@ -17,7 +17,7 @@ Resistor::Resistor(string name, vector<string> args, vector<float> extraInfo)
     types.push_back(componentType::conductanceSource);
 }
 
-Resistor::Resistor(string _name, float r, int n1, int n2)
+Resistor::Resistor(string _name, double r, int n1, int n2)
     :Component{_name}, conductance{1/r}{
 
     nodes.push_back(n1);
