+++ main_project/input/input.hpp
@@ -12,6 +12,6 @@ void readSpice(Circuit& c, string file);
 // this function setsup component c from what it reads from istream
 void readSpice(Circuit& c, istream& file);
 // this function setsup basic variables for circuit
-void setupBasic(Circuit& c, float timeStep);
+void setupBasic(Circuit& c, double timeStep);
 
 #endif
